provisionalmente listo, no me siento satisfecho pero no hay tiepo
