+++ app/src/main/AndroidManifest.xml
@@ -19,14 +19,14 @@
         </activity>
         <activity android:name=".InicioActivity">
         </activity>
-        <activity android:name=".NuevoEmpleadoActivity">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".NuevoEmpleadoActivity">
         </activity>
         <activity android:name=".RegistroActivity">
         </activity>
