+++ app/src/main/res/layout/activity_main.xml
@@ -40,9 +40,9 @@
     <EditText
             android:layout_width="270dp"
             android:layout_height="wrap_content"
-            android:inputType="textPersonName"
+            android:inputType="textEmailAddress"
             android:ems="10"
-            android:id="@+id/txtTelefono" android:layout_marginStart="8dp"
+            android:id="@+id/txtCorreo" android:layout_marginStart="8dp"
             app:layout_constraintStart_toStartOf="parent" android:layout_marginLeft="8dp" android:layout_marginEnd="8dp"
             app:layout_constraintEnd_toEndOf="parent" android:layout_marginRight="8dp"
             android:layout_marginBottom="24dp"
